Remove clj-time.jdbc
See https://github.com/luminus-framework/luminus-template/issues/298
Somewhere in the middleware, org.joda.time.DateTime is not able to be
converted into transit. That bug says it's fixed, but I can't get it to
work with clj-time.jdbc in the project, and I keep having trouble
loading plain JSON responses from the server, so I'm removing
clj-time.jdbc so it doesn't affect the project.