Load main page (single page app) for any relevant requests
This is handled in running.routes.home namespace. Note the repeated routes that all render the home page. We keep the request the same so that bidi can figure out the route once the client has loaded the page.

Also cleaned up some unused or unknown imports in the home page template, and configured production builds to not be in debug mode