WIP heatmap stuff
move heatmap data into {:heatmap-data {:dataset }} so that it is in the default location that rid3 looks for data
update the code that renders each individual day cell into the svg. Keeping a commented-out section that uses the thread-first operator instead of the rid3-> function to do the same result; check out the use of goog.object/get in the anonymous functions.
Update color selection method as well.