WIP heatmap stuff: translated year range and days in year code
generate-dates-for-year: return a js array of every date in a given year
year-range: return a js array of each year between two years
ugly-svg: this code appends a range of svg elements (one for every year between 2003 and 2019), giving each the class "Blues". It then selects those svgs by class name and sets some attributes on them. I wasn't able to get the syntax right to do this in a single step, but this way should also work for creating the rects for every day in each of the years, and then we can fill them in with data.