dereference heatmap-years subscription inside function call
This updates at the right time for the most part. "imagine a graph" title gets correct year, and multiple components are rendered when multiple years are selected, but the heatmap itself isn't filled in unless you navigate to another page and come back, so somewhere we aren't updating correctly

Don't dispatch events for setting year and fetching data on panel load, just use the default stuff.